requirements.txt
# ===== Core Dependencies =====
fastapi==0.115.6  # Updated for Pydantic 2.10 compatibility
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# ===== Database =====
sqlalchemy==2.0.36  # Updated for Python 3.13 compatibility
alembic==1.13.1
psycopg[binary]==3.2.3  # PostgreSQL adapter (v3, compatible with Python 3.13)

# ===== Authentication & Security =====
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # Compatible with passlib 1.7.4

# ===== Validation =====
pydantic==2.10.5  # Updated for Python 3.13 compatibility (has pre-built wheels)
pydantic-settings==2.7.1
email-validator==2.2.0

# ===== Templates & Markdown =====
jinja2==3.1.3
markdown==3.5.2
pymdown-extensions==10.7

# ===== Utilities =====
python-dotenv==1.0.0
